Move shared components to components directory
