+++ blocks/components/editable/index.js
@@ -10,7 +10,7 @@ import 'element-closest';
 /**
  * WordPress dependencies
  */
-import Toolbar from 'editor/components/toolbar';
+import Toolbar from 'components/toolbar';
 
 /**
  * Internal dependencies
