+++ editor/components/inserter/index.js
@@ -4,10 +4,14 @@
 import clickOutside from 'react-click-outside';
 
 /**
+ * WordPress dependencies
+ */
+import IconButton from 'components/icon-button';
+
+/**
  * Internal dependencies
  */
 import InserterMenu from './menu';
-import IconButton from '../icon-button';
 
 class Inserter extends wp.element.Component {
 	constructor() {
