+++ editor/header/saved-state/index.js
@@ -5,10 +5,14 @@ import { connect } from 'react-redux';
 import classNames from 'classnames';
 
 /**
+ * WordPress dependencies
+ */
+import Dashicon from 'components/dashicon';
+
+/**
  * Internal dependencies
  */
 import './style.scss';
-import Dashicon from '../../components/dashicon';
 
 function SavedState( { isDirty } ) {
 	const classes = classNames( 'editor-saved-state', {
